Add new backend checks utils.
The step function to do the backend checks can only handle 500 or so
files before the json that is passed between steps is too large. AWS
recommends passing the data around in S3.

This library publishes the case classes for the json in S3 and two
helper methods to read and write from it.
This way, if we change the json, we only need to change it in one place.

I didn't put these methods in the aws-utils project because the case
classes are very specific to the backend checks.